Moved credentials extraction from AppCoordinator to CredentialsManager
